Fix defaultLocale property to support variant code without country
It was possible to set a default locale with variant code, but only
together with country, which sometime it's not relevant. It was only
possible to set `nl`, `nl_NL_VARIANT`, but it's not possible to set
`nl__VARIANT`.
This fix made a possibility to set locale as `nl__VARIANT`

Also fixed start-env.py script to not require gulp to be installed
globally.